Struct yaml_rust::emitter::YamlEmitter

pub struct YamlEmitter<'a> { /* fields omitted */ }

Methods

impl<'a> YamlEmitter<'a>

pub fn new(writer: &'a mut Write) -> YamlEmitter

pub fn compact(&mut self, compact: bool)

Set 'compact inline notation' on or off, as described for block sequences and mappings.

In this form, blocks cannot have any properties (such as anchors or tags), which should be OK, because this emitter doesn't (currently) emit those anyways.

pub fn is_compact(&self) -> bool

Determine if this emitter is using 'compact inline notation'.

pub fn dump(&mut self, doc: &Yaml) -> EmitResult