Enum yaml_rust::yaml::Yaml

pub enum Yaml {
    Real(String),
    Integer(i64),
    String(String),
    Boolean(bool),
    Array(Array),
    Hash(Hash),
    Alias(usize),
    Null,
    BadValue,
}

A YAML node is stored as this Yaml enumeration, which provides an easy way to access your YAML document.

Examples

use yaml_rust::Yaml;
let foo = Yaml::from_str("-123"); // convert the string to the appropriate YAML type
assert_eq!(foo.as_i64().unwrap(), -123);

// iterate over an Array
let vec = Yaml::Array(vec![Yaml::Integer(1), Yaml::Integer(2)]);
for v in vec.as_vec().unwrap() {
    assert!(v.as_i64().is_some());
}

Variants
Real(String)

Float types are stored as String and parsed on demand. Note that f64 does NOT implement Eq trait and can NOT be stored in BTreeMap.

Integer(i64)

YAML int is stored as i64.

String(String)

YAML scalar.

Boolean(bool)

YAML bool, e.g. true or false.

Array(Array)

YAML array, can be accessed as a Vec.

Hash(Hash)

YAML hash, can be accessed as a LinkedHashMap.

Itertion order will match the order of insertion into the map.

Alias(usize)

Alias, not fully supported yet.

Null

YAML null, e.g. null or ~.

BadValue

Accessing a nonexistent node via the Index trait returns BadValue. This simplifies error handling in the calling code. Invalid type conversion also returns BadValue.

Methods

impl Yaml

pub fn as_bool(&self) -> Option<bool>

pub fn as_i64(&self) -> Option<i64>

pub fn as_str(&self) -> Option<&str>

pub fn as_hash(&self) -> Option<&Hash>

pub fn as_vec(&self) -> Option<&Array>

pub fn into_bool(self) -> Option<bool>

pub fn into_i64(self) -> Option<i64>

pub fn into_string(self) -> Option<String>

pub fn into_hash(self) -> Option<Hash>

pub fn into_vec(self) -> Option<Array>

pub fn is_null(&self) -> bool

pub fn is_badvalue(&self) -> bool

pub fn is_array(&self) -> bool

pub fn as_f64(&self) -> Option<f64>

pub fn into_f64(self) -> Option<f64>

impl Yaml

pub fn from_str(v: &str) -> Yaml