Enum yaml_rust::yaml::Yaml

pub enum Yaml {
    Real(String),
    Integer(i64),
    String(String),
    Boolean(bool),
    Array(Array),
    Hash(Hash),
    Alias(usize),
    Null,
    BadValue,
}

An YAML node is store as this Yaml enumeration, it provides an easy way to access your YAML document.

Examples

use yaml_rust::Yaml;
let foo = Yaml::from_str("-123"); // convert the string to the appropriate YAML type
assert_eq!(foo.as_i64().unwrap(), -123);
 
// iterator over an Array
let vec = Yaml::Array(vec![Yaml::Integer(1), Yaml::Integer(2)]);
for v in vec.as_vec().unwrap() {
    assert!(v.as_i64().is_some());
}

Variants

Real

float types are stored as String, and parsed on demand. Note that f64 does NOT implement Eq trait and can NOT be stored in BTreeMap

Integer

Yaml int is stored as i64.

String

Yaml scalar.

Boolean

Yaml bool, e.g. true or false.

Array

Yaml array, can be access as a Vec.

Hash

Yaml hash, can be access as a BTreeMap.

Alias

Alias, not fully supported yet.

Null

Yaml bool, e.g. null or ~.

BadValue

Access non-exist node by Index trait will return BadValue. This simplifies error handling of user. Invalid type conversion also return BadValue.

Methods

impl Yaml

fn as_bool(&self) -> Option<bool>

fn as_i64(&self) -> Option<i64>

fn as_str(&self) -> Option<&str>

fn as_hash(&self) -> Option<&Hash>

fn as_vec(&self) -> Option<&Array>

fn is_null(&self) -> bool

fn is_badvalue(&self) -> bool

fn as_f64(&self) -> Option<f64>

fn from_str(v: &str) -> Yaml