Enum yaml_rust::yaml::Yaml

pub enum Yaml {
    Real(String),
    Integer(i64),
    String(String),
    Boolean(bool),
    Array(Array),
    Hash(Hash),
    Alias(usize),
    Null,
    BadValue,
}

A YAML node is stored as this Yaml enumeration, which provides an easy way to access your YAML document.

Examples

use yaml_rust::Yaml;
let foo = Yaml::from_str("-123"); // convert the string to the appropriate YAML type
assert_eq!(foo.as_i64().unwrap(), -123);

// iterate over an Array
let vec = Yaml::Array(vec![Yaml::Integer(1), Yaml::Integer(2)]);
for v in vec.as_vec().unwrap() {
    assert!(v.as_i64().is_some());
}

Variants
Real(String)

Float types are stored as String and parsed on demand. Note that f64 does NOT implement Eq trait and can NOT be stored in BTreeMap.

Integer(i64)

YAML int is stored as i64.

String(String)

YAML scalar.

Boolean(bool)

YAML bool, e.g. true or false.

Array(Array)

YAML array, can be accessed as a Vec.

Hash(Hash)

YAML hash, can be accessed as a BTreeMap.

If the order of keys is meaningful, enable the preserve_order feature to store hashes as a LinkedHashMap intead of BTreeMap. When using a LinkedHashMap, the itertion order will match the order of insertion into the map.

yaml-rust = { version = "*", features = ["preserve_order"] }

Alias(usize)

Alias, not fully supported yet.

Null

YAML null, e.g. null or ~.

BadValue

Accessing a nonexistent node via the Index trait returns BadValue. This simplifies error handling in the calling code. Invalid type conversion also returns BadValue.

Methods

impl Yaml

fn as_bool(&self) -> Option<bool>

fn as_i64(&self) -> Option<i64>

fn as_str(&self) -> Option<&str>

fn as_hash(&self) -> Option<&Hash>

fn as_vec(&self) -> Option<&Array>

fn into_bool(self) -> Option<bool>

fn into_i64(self) -> Option<i64>

fn into_string(self) -> Option<String>

fn into_hash(self) -> Option<Hash>

fn into_vec(self) -> Option<Array>

fn is_null(&self) -> bool

fn is_badvalue(&self) -> bool

fn as_f64(&self) -> Option<f64>

fn into_f64(self) -> Option<f64>

impl Yaml

fn from_str(v: &str) -> Yaml